+++ include/config.h
@@ -17,14 +17,22 @@ inline constexpr uint16_t OSC_TARGET_PORT = 9000;
 #endif
 
 inline constexpr char APP_NAME[] = "ChainOSCPad";
-inline constexpr char APP_VERSION[] = "0.6.1";
+inline constexpr char APP_VERSION[] = "0.7.0";
 
 #if defined(CHAINOSCPAD_BOARD_XIAO_ESP32S3)
 inline constexpr char HARDWARE_NAME[] = "XIAO ESP32S3";
+#elif defined(CHAINOSCPAD_BOARD_XIAO_ESP32C3)
+inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C3";
+#elif defined(CHAINOSCPAD_BOARD_XIAO_ESP32C5)
+inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C5";
 #elif defined(CHAINOSCPAD_BOARD_XIAO_ESP32C6)
 inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C6";
 #elif defined(CONFIG_IDF_TARGET_ESP32S3)
 inline constexpr char HARDWARE_NAME[] = "XIAO ESP32S3";
+#elif defined(CONFIG_IDF_TARGET_ESP32C3)
+inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C3";
+#elif defined(CONFIG_IDF_TARGET_ESP32C5)
+inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C5";
 #elif defined(CONFIG_IDF_TARGET_ESP32C6)
 inline constexpr char HARDWARE_NAME[] = "XIAO ESP32C6";
 #else